uhd-firmware : Hacked recipe to build packages from the master release area.
I hate this. HATE HATE HATE.

Insert a DEFAULT_PREFERENCE = "-1" in this recipe after the next firmware
release. Better yet, delete this recipe and never build images from
unversioned firmware.

This recipe sucks. It exists only to create images when Josh won't create
a tag.
THE PV is basically wrong at this point. We use PR to make this package
install instead of a firmware package built from a release.

Signed-off-by: Philip Balister <philip@opensdr.com>